Centralize Python module initialization logic
This change introduces a new Tf.PreparePythonModule function
that serves as a central location for the Python module
initialization code that was previously copied to all of
the various __init__.py files in the repository. This
allows us to get rid of a bunch of boiler-plate code
and also provides a single place to add more logic in
the future. In particular, this may be useful for pending
work to add support for Python 3.8+ on Windows.

See #1404

(Internal change: 2161883)